Add is_followed field to UserResponseType and ProfileType
This commit adds the is_followed field to the UserResponseType and ProfileType in the client/src/@types/user.ts file. This field is used to indicate whether a user is followed or not.
Refactor middleware in settings.py

This commit refactors the middleware in the server/communiche/settings.py file. The corsheaders.middleware.CorsMiddleware is now placed before other middlewares to ensure proper handling of CORS requests.

Update URLs in urls.py

This commit updates the URLs in the server/communiche/urls.py file. Two new paths, follow/<int:user_id>/<int:follower_id> and unfollow/<int:user_id>/<int:follower_id>, are added to handle user follow and unfollow requests.

Add UserFollowing model

This commit adds the UserFollowing model to the server/communiche/models.py file. This model represents the relationship between users where one user follows another. It includes fields for the follower, following, and created_at timestamp.

Add followUser and unfollowUser functions to UserService.ts

This commit adds the followUser and unfollowUser functions to the client/src/services/UserService.ts file. These functions are used to send requests to the server to follow or unfollow a user.

Update ApiService.ts

This commit updates the ApiService.ts file in the client/src/services directory. It adds the Authorization header to the request config if a token is available in the local storage. This ensures that authenticated requests include the necessary authorization information.

Add UserFollowingSerializer to serializers.py

This commit adds the UserFollowingSerializer to the server/communiche/serializers.py file. This serializer is used to serialize the UserFollowing model and includes fields for the follower, following, and created_at timestamp.

Add FollowButton component

This commit adds the FollowButton component to the client/src/views/profile/FollowButton.tsx file. This component is used to display a button that allows users to follow or unfollow another user. It handles the follow/unfollow logic and displays an error message if an error occurs.